refactor(deps): relocate 4 shared leaf helpers to internal/ (PR-D1, decouple prep)
Verbatim leaf relocations so the rotation engine + the soon-to-move console
supervisor can reach helpers the core CLI also keeps (PR-B-slice-1 style:
heterogeneous leaf moves, one PR). No behavior change.

- S3 helpers parseS3URL/newS3Client/buildS3Key/s3ObjectExists/uploadFile →
  internal/storage (ParseS3URL/NewS3Client/BuildS3Key/S3ObjectExists/UploadFile);
  rewire archive/baseline/upload/rotate. rotate's uploadFileFunc test seam now
  points at storage.UploadFile (stays in cmd until the engine moves in PR-D2).
- parseRetain → cliutil.ParseRetain (generic Nd/Nh duration parser; also used
  by agent + doctor).
- scrubMonitorErr/scrubMonitorErrText → config.ScrubDSNError/ScrubDSNText
  (both surviving consumers — the supervisor and the up-rotation loop — already
  import config; the loop needs it before it moves to internal/rotation).
- streamDeps → internal/streamdeps.Default() (kept OUT of streamrun to preserve
  PR-A's Deps DI seam); rewire stream + monitor.

Tests TestParseRetain*/TestParseS3URL* moved to their new homes. build/vet
(unit+integration tags) + full unit suite green; core diff limited to the
rewired cmd files + the new internal packages.
